+++ RELEASE_NOTES.md
@@ -10,6 +10,12 @@ v1.0.0-beta.x.x
   in Python.
   [#1186](https://github.com/OpenAssetIO/OpenAssetIO/issues/1186)
 
+### Bug fixes
+
+- Throw an exception when attempting to copy-construct a null
+  `TraitsData`.
+  [#1153](https://github.com/OpenAssetIO/OpenAssetIO/issues/1153)
+
 v1.0.0-beta.1.0
 ---------------